backend/requirements.txt
# FastAPI ve sunucu
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Veritabanı
sqlalchemy==2.0.23
alembic==1.12.1

# Text extraction ve OCR
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Yardımcı kütüphaneler
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0